requirements.txt
fastapi
uvicorn
sqlalchemy
python-telegram-bot
pydantic
requests
